requirements.txt
python-dotenv>=1.0.0
livekit>=0.17.0
livekit-agents>=0.9.0
livekit-plugins-assemblyai>=0.7.0
livekit-plugins-openai>=0.8.0
livekit-plugins-cartesia>=0.3.0
livekit-plugins-silero>=0.7.0
livekit-plugins-noise-cancellation>=0.1.0
livekit-plugins-turn-detector>=0.1.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
twilio>=9.2.3
python-multipart>=0.0.9
google-genai
